+++ src/Aggregator/Infrastructure/Crawler/Source/Data/MediaCongoNet.php
@@ -30,7 +30,7 @@ final class MediaCongoNet extends Source
     public function fetch(FetchConfig $config): void
     {
         $this->initialize();
-        $page = $config->page ?? PageRange::from(sprintf('1:%d', $this->getLastPage(self::URL . '/articles.html')));
+        $page = $config->page ?? $this->getPagination();
 
         for ($i = $page->start; $i < $page->end; $i++) {
             try {
@@ -74,9 +74,15 @@ final class MediaCongoNet extends Source
                 $this->skip($interval, $timestamp, $title, $date);
             }
         } catch (\Throwable $e) {
-            $this->logger->critical("> {$e->getMessage()} [Failed] ❌");
+            $this->logger->error("> {$e->getMessage()} [Failed] ❌");
             return;
         }
+    }
+
+    #[\Override]
+    public function getPagination(?string $category = null): PageRange
+    {
+        return PageRange::from(sprintf('1:%d', $this->getLastPage(self::URL . '/articles.html')));
     }
 
     #[\Override]