+++ src/Aggregator/Infrastructure/Crawler/Source/Data/RadioOkapiNet.php
@@ -28,7 +28,7 @@ final class RadioOkapiNet extends Source
     public function fetch(FetchConfig $config): void
     {
         $this->initialize();
-        $page = $config->page ?? PageRange::from(sprintf('0:%d', $this->getLastPage(self::URL . '/actualite')));
+        $page = $config->page ?? $this->getPagination();
 
         for ($i = $page->start; $i <= $page->end; $i++) {
             try {
@@ -73,8 +73,14 @@ final class RadioOkapiNet extends Source
                 $this->skip($interval, $timestamp, $title, $date);
             }
         } catch (\Throwable $e) {
-            $this->logger->critical("> {$e->getMessage()} [Failed] ❌");
+            $this->logger->error("> {$e->getMessage()} [Failed] ❌");
             return;
         }
+    }
+
+    #[\Override]
+    public function getPagination(?string $category = null): PageRange
+    {
+        return PageRange::from(sprintf('0:%d', $this->getLastPage(self::URL . '/actualite')));
     }
 }